+++ template/src/hooks/use-set-query-online-manager.ts
@@ -1,0 +1,11 @@
+import { useEffect } from 'react';
+import NetInfo from '@react-native-community/netinfo';
+import { onlineManager } from '@tanstack/react-query';
+
+export function useSetQueryOnlineManager() {
+  useEffect(() => {
+    return NetInfo.addEventListener((state) => {
+      onlineManager.setOnline(!!state.isConnected);
+    });
+  }, []);
+}
